+++ src/User.php
@@ -4294,46 +4294,6 @@ JAVASCRIPT;
 
 
     /**
-     * Get all groups where the current user have delegating.
-     *
-     * @since 0.83
-     *
-     * @param integer|string $entities_id ID of the entity to restrict
-     *
-     * @return integer[]
-     */
-    public static function getDelegateGroupsForUser($entities_id = '')
-    {
-        /** @var \DBmysql $DB */
-        global $DB;
-
-        $iterator = $DB->request([
-            'SELECT'          => 'glpi_groups_users.groups_id',
-            'DISTINCT'        => true,
-            'FROM'            => 'glpi_groups_users',
-            'INNER JOIN'      => [
-                'glpi_groups'  => [
-                    'FKEY'   => [
-                        'glpi_groups_users'  => 'groups_id',
-                        'glpi_groups'        => 'id'
-                    ]
-                ]
-            ],
-            'WHERE'           => [
-                'glpi_groups_users.users_id'        => Session::getLoginUserID(),
-                'glpi_groups_users.is_userdelegate' => 1
-            ] + getEntitiesRestrictCriteria('glpi_groups', '', $entities_id, 1)
-        ]);
-
-        $groups = [];
-        foreach ($iterator as $data) {
-            $groups[$data['groups_id']] = $data['groups_id'];
-        }
-        return $groups;
-    }
-
-
-    /**
      * Execute the query to select box with all glpi users where select key = name
      *
      * Internaly used by showGroup_Users, dropdownUsers and ajax/getDropdownUsers.php
@@ -4387,42 +4347,6 @@ JAVASCRIPT;
 
             case "id":
                 $WHERE = ['glpi_users.id' => Session::getLoginUserID()];
-                break;
-
-            case "delegate":
-                $groups = self::getDelegateGroupsForUser($entity_restrict);
-                $users  = [];
-                if (count($groups)) {
-                    $iterator = $DB->request([
-                        'SELECT'    => 'glpi_users.id',
-                        'FROM'      => 'glpi_groups_users',
-                        'LEFT JOIN' => [
-                            'glpi_users'   => [
-                                'FKEY'   => [
-                                    'glpi_groups_users'  => 'users_id',
-                                    'glpi_users'         => 'id'
-                                ]
-                            ]
-                        ],
-                        'WHERE'     => [
-                            'glpi_groups_users.groups_id' => $groups,
-                            'glpi_groups_users.users_id'  => ['<>', Session::getLoginUserID()]
-                        ]
-                    ]);
-                    foreach ($iterator as $data) {
-                           $users[$data["id"]] = $data["id"];
-                    }
-                }
-               // Add me to users list for central
-                if (Session::getCurrentInterface() == 'central') {
-                    $users[Session::getLoginUserID()] = Session::getLoginUserID();
-                }
-
-                if (count($users)) {
-                    $WHERE = ['glpi_users.id' => $users];
-                } else {
-                    $WHERE = ['0'];
-                }
                 break;
 
             case "groups":